+++ guillotina/utils.py
@@ -2,7 +2,6 @@
 from aiohttp.web_exceptions import HTTPUnauthorized
 from collections import MutableMapping
 from hashlib import sha256 as sha
-from zope.dottedname.resolve import resolve
 
 import fnmatch
 import importlib
@@ -135,6 +134,31 @@ class Lazy(object):
         inst.__dict__[name] = value
 
         return value
+
+
+def resolve(name, module=None):
+    name = name.split('.')
+    if not name[0]:
+        if module is None:
+            raise ValueError("relative name without base module")
+        module = module.split('.')
+        name.pop(0)
+        while not name[0]:
+            module.pop()
+            name.pop(0)
+        name = module + name
+
+    used = name.pop(0)
+    found = __import__(used)
+    for n in name:
+        used += '.' + n
+        try:
+            found = getattr(found, n)
+        except AttributeError:
+            __import__(used)
+            found = getattr(found, n)
+
+    return found
 
 
 def resolve_or_get(potential_dotted_name):